+++ CHANGELOG.md
@@ -14,6 +14,8 @@ NOTE: Breaking release in controllers.
 - Minimum Go version v1.13 (error wrapping required).
 - Refactor Logger with structured logging.
 - Add Logrus helper wrapper.
+- Refactor to simplify the retrievers.
+- Add multiretriever to retriever different resource types on the same controller.
 
 ## [0.8.0] - 2019-12-11